rf-appium/pyproject.toml
[project]
name = "rf-appium"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.0.1",
    "robotframework>=7.2.2",
    "robotframework-appiumlibrary>=2.1.0",
]
